Finished all the yellow problems
