+++ src/pages/ContactUs.jsx
@@ -4,23 +4,27 @@ import Navbar from '../components/Navbar';
 import Footer from '../components/Footer';
 
 const Container = styled.div`
-  
+  background-color: #d6c9c9;
 `
-
 const ContactUsPageWrapper = styled.div`
+  background-color: #fff;
   display: flex;
   flex-direction: column;
   align-items: center;
-  margin-top: 50px;
-`;
+  background-color: #fff;
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 40px;
+`
+const Heading = styled.h1`
 
+`
 const Form = styled.form`
   display: flex;
   flex-direction: column;
   align-items: center;
   width: 400px;
   background-color: #fff;
-  /* border: 1px solid #ccc; */
   border-radius: 5px;
   padding: 20px;
 `;
@@ -29,6 +33,9 @@ const Input = styled.input`
   padding: 10px;
   margin: 10px 0;
   border-radius: 5px;
+  background-color: #75ffbf;
+  font-weight: 500;
+  font-family: Arial, Helvetica, sans-serif;
   border: 1px solid #ccc;
   width: 100%;
 `;
@@ -38,6 +45,9 @@ const TextArea = styled.textarea`
   margin: 10px 0;
   border-radius: 5px;
   border: 1px solid #ccc;
+  background-color: #75ffba;
+  font-weight: 400;
+  font-family: Arial, Helvetica, sans-serif;
   width: 100%;
   height: 150px;
 `;
@@ -45,7 +55,7 @@ const TextArea = styled.textarea`
 const Button = styled.button`
   padding: 10px 20px;
   margin-top: 20px;
-  background-color: #0077ff;
+  background-color: #000000;
   color: #fff;
   border: none;
   border-radius: 5px;
@@ -66,7 +76,7 @@ const ContactUs = () => {
     <Container>
     <Navbar/>
     <ContactUsPageWrapper>
-      <h2>Contact Us</h2>
+      <Heading><hr/>Contact Us<hr/></Heading>
       <Form onSubmit={handleSubmit}>
         <Input
           type="text"
